// src/dorc-web/src/apis/dorc-api/models/GetRefDataAuditListResponseDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type {
    RefDataAuditApiModel,
} from './';

/**
 * @export
 * @interface GetRefDataAuditListResponseDto
 */
export interface GetRefDataAuditListResponseDto {
    /**
     * @type {number}
     * @memberof GetRefDataAuditListResponseDto
     */
    CurrentPage?: number;
    /**
     * @type {number}
     * @memberof GetRefDataAuditListResponseDto
     */
    TotalItems?: number;
    /**
     * @type {number}
     * @memberof GetRefDataAuditListResponseDto
     */
    TotalPages?: number;
    /**
     * @type {Array<RefDataAuditApiModel>}
     * @memberof GetRefDataAuditListResponseDto
     */
    Items?: Array<RefDataAuditApiModel> | null;
}